+++ docs/src/SelectList.doc.js
@@ -24,8 +24,14 @@ card(
       },
       {
         name: 'errorMessage',
-        type: '?string',
+        type: 'string',
         href: 'exampleWithError',
+      },
+      {
+        name: 'helperText',
+        type: 'string',
+        description: 'More information about how to complete the form field',
+        href: 'helperText',
       },
       {
         name: 'id',
@@ -34,8 +40,12 @@ card(
         href: 'basicExample',
       },
       {
+        name: 'label',
+        type: 'string',
+      },
+      {
         name: 'name',
-        type: '?string',
+        type: 'string',
         href: 'basicExample',
       },
       {
@@ -52,12 +62,12 @@ card(
       },
       {
         name: 'placeholder',
-        type: '?string',
+        type: 'string',
         href: 'basicExample',
       },
       {
         name: 'value',
-        type: '?string',
+        type: 'string',
         description: 'Value that is selected.',
         href: 'basicExample',
       },
@@ -71,56 +81,76 @@ card(
     name="Example"
     description={`Make sure to attach a \`Label\` to every SelectList.`}
     defaultCode={`
-class SelectListExample extends React.Component {
-  constructor(props) {
-    super(props);
-    this.handleChange = this._handleChange.bind(this);
-    this.state = {
-      city: 'la'
+function Example(props) {
+  const [city, setCity] = React.useState('la')
+  const cityOptions = [
+    {
+      value: "bos",
+      label: "Boston"
+    },
+    {
+      value: "la",
+      label: "Los Angeles"
+    },
+    {
+      value: "sf",
+      label: "San Francisco"
     }
-  }
-
-  _handleChange({ value }) {
-    this.setState({
-      city: value,
-    })
-  }
-
-  render() {
-    const cityOptions = [
-      {
-        value: "bos",
-        label: "Boston"
-      },
-      {
-        value: "la",
-        label: "Los Angeles"
-      },
-      {
-        value: "sf",
-        label: "San Francisco"
-      }
-    ];
-    return (
-      <Box>
-        <Box paddingY={2}>
-          <Label htmlFor="city">
-            <Text>City</Text>
-          </Label>
-        </Box>
-        <SelectList
-          id="city"
-          name="city"
-          onChange={this.handleChange}
-          options={cityOptions}
-          placeholder="Select city"
-          value={this.state.city}
-        />
-      </Box>
-    );
-  }
+  ];
+  return (
+    <SelectList
+      id="city"
+      name="city"
+      onChange={({ value }) => setCity(value)}
+      options={cityOptions}
+      placeholder="Select city"
+      label="City"
+      value={city}
+    />
+  );
 }
     `}
+  />
+);
+
+card(
+  <Example
+    id="helperText"
+    name="Example: Helper Text"
+    description={`Whenever you want to provide more information about a form field, you should use \`helperText\`.`}
+    defaultCode={`
+function Example(props) {
+  const [country, setCountry] = React.useState('belgium')
+  const countries = [
+    {
+      value: "belgium",
+      label: "Belgium"
+    },
+    {
+      value: "france",
+      label: "France"
+    },
+    {
+      value: "usa",
+      label: "USA"
+    }
+  ];
+  return (
+    <Box padding={2} color="white">
+      <SelectList
+        id="country"
+        name="country"
+        onChange={({ value }) => setCountry(value)}
+        options={countries}
+        placeholder="Select country"
+        label="Country"
+        helperText="Pick your favourite country (the one with the best fries)"
+        value={country}
+      />
+    </Box>
+  );
+}
+`}
   />
 );
 
@@ -129,58 +159,36 @@ card(
     id="exampleWithError"
     name="Example: With Error Message"
     description={`SelectList can display error messages if you'd like.
-    To use our errors, simply pass in an \`errorMessage\` when there is an error present and we will
-    handle the rest.`}
+    To use our errors, simply pass in an \`errorMessage\` when there is an error present.`}
     defaultCode={`
-class SelectListExample extends React.Component {
-  constructor(props) {
-    super(props);
-    this.handleChange = this._handleChange.bind(this);
-    this.state = {
-      city: 'la'
+function Example(props) {
+  const [hobby, setHobby] = React.useState('sailing')
+  const hobbies = [
+    {
+      value: "sailing",
+      label: "Sailing"
+    },
+    {
+      value: "skiing",
+      label: "Skiing"
+    },
+    {
+      value: "cycling",
+      label: "Cycling"
     }
-  }
-
-  _handleChange({ value }) {
-    this.setState({
-      city: value,
-    })
-  }
-
-  render() {
-    const cityOptions = [
-      {
-        value: "bos",
-        label: "Boston"
-      },
-      {
-        value: "la",
-        label: "Los Angeles"
-      },
-      {
-        value: "sf",
-        label: "San Francisco"
-      }
-    ];
-    return (
-      <Box>
-        <Box paddingY={2}>
-          <Label htmlFor="city">
-            <Text>City</Text>
-          </Label>
-        </Box>
-        <SelectList
-          id="city"
-          name="city"
-          errorMessage="This selection has an error"
-          onChange={this.handleChange}
-          options={cityOptions}
-          placeholder="Select city"
-          value={this.state.city}
-        />
-      </Box>
-    );
-  }
+  ];
+  return (
+    <SelectList
+      id="hobby"
+      name="hobby"
+      errorMessage={ hobby === 'sailing' ? 'Sailing is not a hobby, it is a passion' : null}
+      onChange={({ value }) => setHobby(value)}
+      options={hobbies}
+      placeholder="Select hobby"
+      label="Hobby"
+      value={hobby}
+    />
+  );
 }
     `}
   />