+++ packages/gestalt/src/FormErrorMessage.js
@@ -2,18 +2,22 @@
 
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import Box from './Box.js';
 import Text from './Text.js';
 
-type Props = {|
+export default function FormErrorMessage({
+  id,
+  text = '',
+}: {|
   id: string,
   text?: string,
-|};
-
-export default function FormErrorMessage({ id, text = '' }: Props) {
+|}) {
   return (
-    <Text color="red">
-      <span id={`${id}-error`}>{text}</span>
-    </Text>
+    <Box marginTop={2} paddingX={2}>
+      <Text color="red" size="sm">
+        <span id={`${id}-error`}>{text}</span>
+      </Text>
+    </Box>
   );
 }