+++ packages/gestalt/src/Text.js
@@ -8,11 +8,10 @@ import styles from './Text.css';
 import typography from './Typography.css';
 
 const SIZE_SCALE: { [size: ?string]: number } = {
-  xs: 1,
-  sm: 2,
-  md: 3,
-  lg: 4,
-  xl: 5,
+  sm: 1,
+  md: 2,
+  lg: 3,
+  xl: 4,
 };
 
 type Props = {|
@@ -39,10 +38,7 @@ type Props = {|
   inline?: boolean,
   italic?: boolean,
   overflow?: 'normal' | 'breakWord',
-  size?: 'xs' | 'sm' | 'md' | 'lg' | 'xl',
-  smSize?: 'xs' | 'sm' | 'md' | 'lg' | 'xl',
-  mdSize?: 'xs' | 'sm' | 'md' | 'lg' | 'xl',
-  lgSize?: 'xs' | 'sm' | 'md' | 'lg' | 'xl',
+  size?: 'sm' | 'md' | 'lg' | 'xl', // xl is intentionally undocumented because it is depreacted
   leading?: 'tall' | 'short',
   truncate?: boolean,
   weight?: 'bold' | 'normal',
@@ -56,26 +52,17 @@ export default function Text({
   inline = false,
   italic = false,
   overflow = 'breakWord',
-  size = 'md',
-  smSize,
-  mdSize,
-  lgSize,
+  size = 'lg',
   leading,
   truncate = false,
   weight = 'normal',
   __dangerouslyIncreaseLineHeight = false,
 }: Props) {
   const scale = SIZE_SCALE[size];
-  const smScale = SIZE_SCALE[smSize];
-  const mdScale = SIZE_SCALE[mdSize];
-  const lgScale = SIZE_SCALE[lgSize];
 
   const cs = cx(
     styles.Text,
     styles[`fontSize${scale}`],
-    lgSize && styles[`lgFontSize${lgScale}`],
-    mdSize && styles[`mdFontSize${mdScale}`],
-    smSize && styles[`smFontSize${smScale}`],
     color === 'blue' && colors.blue,
     color === 'darkGray' && colors.darkGray,
     color === 'eggplant' && colors.eggplant,
@@ -146,11 +133,8 @@ Text.propTypes = {
   inline: PropTypes.bool,
   italic: PropTypes.bool,
   leading: PropTypes.oneOf(['tall', 'short']),
-  lgSize: PropTypes.oneOf(['xs', 'sm', 'md', 'lg', 'xl']),
-  mdSize: PropTypes.oneOf(['xs', 'sm', 'md', 'lg', 'xl']),
   overflow: PropTypes.oneOf(['normal', 'breakWord']),
-  size: PropTypes.oneOf(['xs', 'sm', 'md', 'lg', 'xl']),
-  smSize: PropTypes.oneOf(['xs', 'sm', 'md', 'lg', 'xl']),
+  size: PropTypes.oneOf(['sm', 'md', 'lg', 'xl']),
   truncate: PropTypes.bool,
   weight: PropTypes.oneOf(['bold', 'normal']),
 };