+++ packages/gestalt/src/TextField.jsdom.test.js
@@ -84,4 +84,47 @@ describe('TextField', () => {
     });
     expect(mockKeyDown).toHaveBeenCalled();
   });
+
+  it('shows a label for the text field', () => {
+    const { getByText } = render(
+      <TextField
+        id="test"
+        label="Label for the text field"
+        onChange={() => {}}
+        value="TextField Text"
+      />
+    );
+    expect(getByText('Label for the text field')).toBeVisible();
+  });
+
+  it('shows helper text for the text field', () => {
+    const { getByText } = render(
+      <TextField
+        id="test"
+        label="Label for the text field"
+        helperText="Helper text for the text field"
+        onChange={() => {}}
+        value="TextField Text"
+      />
+    );
+    expect(getByText('Helper text for the text field')).toBeVisible();
+  });
+
+  it('hides the helper text for the text field when an error message is shown', () => {
+    const { getByText } = render(
+      <TextField
+        id="test"
+        label="Label for the text field"
+        helperText="Helper text for the text field"
+        errorMessage="Error message for the text field"
+        onChange={() => {}}
+        value="TextField Text"
+      />
+    );
+    expect(() => {
+      getByText('Helper text for the text field');
+    }).toThrow(
+      'Unable to find an element with the text: Helper text for the text field'
+    );
+  });
 });