+++ packages/gestalt/src/Tooltip.js
@@ -4,62 +4,125 @@ import * as React from 'react';
 import Controller from './Controller.js';
 import Text from './Text.js';
 import Box from './Box.js';
+import Layer from './Layer.js';
 
 const noop = () => {};
+const TIMEOUT = 100;
 
 type Props = {|
   children: React.Node,
+  link?: React.Node,
+  idealDirection?: 'up' | 'right' | 'down' | 'left',
   inline?: boolean,
   text: string,
 |};
 
-type State = {|
-  hovered: boolean,
-|};
+const initialState = { hoveredIcon: false, hoveredText: false, isOpen: false };
 
-export default class Tooltip extends React.Component<Props, State> {
-  state = {
-    hovered: false,
+const reducer = (state, action) => {
+  switch (action.type) {
+    case 'hoverInIcon':
+      return {
+        ...state,
+        hoveredIcon: true,
+        isOpen: true,
+      };
+    case 'hoverInText':
+      return {
+        ...state,
+        hoveredText: true,
+        isOpen: true,
+      };
+    case 'hoverOutIcon':
+      return {
+        ...state,
+        hoveredIcon: false,
+        ...(!state.hoveredText ? { isOpen: false } : {}),
+      };
+    case 'hoverOutText':
+      return {
+        ...state,
+        hoveredText: false,
+        ...(!state.hoveredIcon ? { isOpen: false } : {}),
+      };
+    default:
+      throw new Error();
+  }
+};
+
+export default function Tooltip({
+  children,
+  link,
+  idealDirection = 'down',
+  inline,
+  text,
+}: Props) {
+  const [state, dispatch] = React.useReducer(reducer, initialState);
+  const { isOpen } = state;
+
+  const childRef = React.useRef<?HTMLDivElement>(null);
+  const { current: anchor } = childRef;
+
+  const handleIconMouseEnter = () => {
+    dispatch({ type: 'hoverInIcon' });
   };
 
-  childRef: {| current: null | React.ElementRef<'div'> |} = React.createRef();
+  const handleIconMouseLeave = () => {
+    setTimeout(() => {
+      dispatch({ type: 'hoverOutIcon' });
+    }, TIMEOUT);
+  };
 
-  handleMouseEnter = () => this.setState({ hovered: true });
+  const handleTextMouseEnter = () => {
+    dispatch({ type: 'hoverInText' });
+  };
 
-  handleMouseLeave = () => this.setState({ hovered: false });
+  const handleTextMouseLeave = () => {
+    setTimeout(() => {
+      dispatch({ type: 'hoverOutText' });
+    }, TIMEOUT);
+  };
 
-  render() {
-    const { children, inline, text } = this.props;
-    const { hovered } = this.state;
-    const { current: anchor } = this.childRef;
-
-    return (
-      <Box display={inline ? 'inlineBlock' : 'block'}>
-        <Box
-          onMouseEnter={this.handleMouseEnter}
-          onMouseLeave={this.handleMouseLeave}
-          ref={this.childRef}
-        >
-          {children}
-        </Box>
-        {hovered && !!anchor && (
+  return (
+    <Box display={inline ? 'inlineBlock' : 'block'}>
+      <Box
+        aria-label={text}
+        ref={childRef}
+        onMouseEnter={handleIconMouseEnter}
+        onMouseLeave={handleIconMouseLeave}
+      >
+        {children}
+      </Box>
+      {isOpen && !!anchor && (
+        <Layer>
           <Controller
             anchor={anchor}
-            bgColor="darkGray"
             caret={false}
-            idealDirection="down"
+            bgColor="darkGray"
+            idealDirection={idealDirection}
             onDismiss={noop}
-            positionRelativeToAnchor
+            positionRelativeToAnchor={false}
             size={null}
           >
-            <Box maxWidth={180} paddingY={1} paddingX={2} role="tooltip">
-              <Text color="white" size="xs">
+            <Box
+              maxWidth={180}
+              paddingY={1}
+              paddingX={2}
+              onBlur={link ? handleTextMouseLeave : undefined}
+              onFocus={link ? handleTextMouseEnter : undefined}
+              onMouseEnter={link ? handleTextMouseEnter : undefined}
+              onMouseLeave={link ? handleTextMouseLeave : undefined}
+              role="tooltip"
+              tabIndex={0}
+            >
+              <Text color="white" size="sm">
                 {text}
               </Text>
+              {link && <Box paddingY={1}>{link}</Box>}
             </Box>
           </Controller>
-        )}
-      </Box>
-    );
-  }
+        </Layer>
+      )}
+    </Box>
+  );
 }